backend/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
joblib>=1.3.0
groq>=0.4.0
python-dotenv>=1.0.0
pymongo>=4.6.0
bcrypt>=4.1.0
PyJWT>=2.8.0
